Remove need for `sat` in bootstrap alg params
TODO: still need to consider how we make sure expected transformations occur for
the datasets they should be applied for. One idea is to add the regression
coefficients to the `BootstrapParams` as optional params. If they're defined,
the bootstrap algorithm can automatically apply them.

This still puts the responsibility on the bootstrap alg itself though. It
shouldn't really need to do any transforms to the data provided to it.